% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiaovbay.R
\name{Multiaovbayes}
\alias{Multiaovbayes}
\title{Interactive panel Two-way ANOVA classic, non parametric and bayesian}
\usage{
Multiaovbayes(dataset = FALSE)
}
\arguments{
\item{dataset}{Data set}
}
\value{
A shiny panel with the classical, non-parametric and Bayesian analyzes of variance, based on the specification of the dependent and independent variable of the data set provided in \code{dataset}, also provides a decision diagram that suggests which method is appropriate, based on the assumptions of the models.
}
\description{
Interactive panel to visualize and develop two-way analysis of variance models, from the classical, non-parametric and Bayesian approach.
}
\examples{
\donttest{
data(ToothGrowth)
ToothGrowth$dose = factor(ToothGrowth$dose)
levels(ToothGrowth$dose) = c("Low", "Medium", "High")
Multiaovbayes(ToothGrowth)
}
}
